Purchases and Details for their Items
The Homescreen shows your purchases and you can click on the purchases to view all the items that were bought. You can also click on the items to see a detailed view of them.

The ItemChangeDialoge was changed to have a variable if the item should be editable or not. This dialoge is now used by the items page and the home page. On the items page the items are editable on the home page not.