Purchases Routes refactoring
Changed the purchases get route to fetch purchases. If you provide a start and end date you get only the purchases that were bought in that timeperiode. If you do not give a start date it return the purchases of the last 28 weeks and if you dont provide a end date it takes the current date.