Shopping Page Change
Shopping Page now shows all Items and has a Search feature. You can ether search by typing in the name of a product or scan the Barcode. Additionally the endpoint for products was changed. It is now possible to search for barcodes there. The user can also scan a barcode to create a new product if he desires.