New Item Creation in Purchase + small Bug Fix
A new dialog opens when you scan a barcode of an item thats not yet part of the database. You can cancel, re-scan a barcode or create a new item that gets directly put into your purchase. Small bug fix were it was possible to change your first and lastname and email to nothing.